# docs/api/openapi.yaml
openapi: 3.1.0
info:
  version: 2.18.5
  title: API GEO Paraguay
  description: A RESTful service that provides detailed geographical information for Paraguay based on coordinates (longitude and latitude).
  contact:
    email: josego85@gmail.com
servers:
  - url: /api/v1
    description: Development server
...
paths:
...
  /departments/{id}:
    get:
      tags:
        - Department
      operationId: getDepartmentById
      summary: Get a department by ID
      description: Retrieves details for a specific department identified by its ID.
      parameters:
        - in: path
          name: id
          description: Department ID
          required: true
          schema:
            type: number
      responses:
        '200':
          description: Successful retrieval of department details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '404':
          description: Department not found
        '500':
          description: Internal server error
...
components:
...
  schemas:
    Department:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the department
        name:
          type: string
          description: Name of the department
        capital_name:
          type: string
          description: Name of the capital city of the department
...